+++ external/vulkancts/modules/vulkan/draw/vktDrawIndexedTest.cpp
@@ -168,7 +168,7 @@ tcu::TestStatus DrawIndexed::iterate (void)
 	submitCommandsAndWait(m_vk, device, queue, m_cmdBuffer.get());
 
 	// Validation
-	tcu::Texture2D referenceFrame(vk::mapVkFormat(m_colorAttachmentFormat), (int)(0.5 + WIDTH), (int)(0.5 + HEIGHT));
+	tcu::Texture2D referenceFrame(vk::mapVkFormat(m_colorAttachmentFormat), (int)(0.5f + static_cast<float>(WIDTH)), (int)(0.5f + static_cast<float>(HEIGHT)));
 								  referenceFrame.allocLevel(0);
 
 	const deInt32 frameWidth	= referenceFrame.getWidth();
@@ -275,7 +275,7 @@ tcu::TestStatus DrawInstancedIndexed::iterate (void)
 	// Validation
 	VK_CHECK(m_vk.queueWaitIdle(queue));
 
-	tcu::Texture2D referenceFrame(vk::mapVkFormat(m_colorAttachmentFormat), (int)(0.5 + WIDTH), (int)(0.5 + HEIGHT));
+	tcu::Texture2D referenceFrame(vk::mapVkFormat(m_colorAttachmentFormat), (int)(0.5f + static_cast<float>(WIDTH)), (int)(0.5f + static_cast<float>(HEIGHT)));
 	referenceFrame.allocLevel(0);
 
 	const deInt32 frameWidth = referenceFrame.getWidth();
